requirements.txt
# pinecone
# streamlit>=1.24.0
# chromadb>=0.4.0
# langchain>=0.1.0
# langchain-community>=0.0.10
# openai>=1.0.0
# python-dotenv>=1.0.0
# tiktoken>=0.5.1
# faiss-cpu>=1.7.4
# sentence-transformers>=2.2.2
# transformers>=4.30.0
# torch>=2.0.0
# pypdf2>=3.0.0
# redis>=4.5.0  
# rank-bm25>=0.2.2 
# asyncio>=3.4.3
# boto3>=1.28.0
# pysqlite3-binary>=0.5.2
# streamlit-extras
# pymysql==1.1.0
# sqlalchemy==2.0.27
# langchain-openai


# requirements.txt - Updated for GraphRAG

# Existing dependencies
streamlit>=1.28.0
openai>=1.3.0
pinecone-client>=3.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
PyPDF2>=3.0.1
pdfplumber>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.2

# New GraphRAG dependencies
neo4j>=5.14.0
spacy>=3.7.0
networkx>=3.2.0
community>=1.0.0b1
python-louvain>=0.16

# Additional NLP and graph analysis
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Performance and caching
redis>=5.0.0
celery>=5.3.0
joblib>=1.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
